refactor: convert markdown components to ts
