+++ src/components/markdown/root-wrapper.tsx
@@ -1,0 +1,103 @@
+import React from 'react';
+
+import Collapsable from '../collapsable';
+import {HeadingTag} from '../heading-tags';
+
+import {MdHeadingTagAnchorDisabled} from './heading-tags';
+
+/**
+ * Determine heading level for a given React element.
+ *
+ * @param node - React element representing a heading.
+ * @returns Heading level (1-6) or -1 if the element is not a heading.
+ */
+export function getHeadingLevel(node: React.ReactElement): number {
+  if (
+    node.type === MdHeadingTagAnchorDisabled ||
+    node.type === HeadingTag
+  ) {
+    return (node.props as any).level;
+  }
+  else if (['h1', 'h2', 'h3', 'h4', 'h5', 'h6'].includes(node.type as any)) {
+    return parseInt((node.type as string).slice(1, 2));
+  }
+  return -1;
+}
+
+/**
+ * Group sequential nodes into collapsable sections based on heading levels.
+ *
+ * @param nodes - Array of markdown child nodes.
+ * @param startIdx - Starting index for processing.
+ * @param minLevel - Minimum heading level considered as a new section.
+ * @returns Tuple containing grouped section elements and the end index.
+ */
+export function groupSections(
+  nodes: React.ReactElement[],
+  startIdx = 0,
+  minLevel = 1
+): [React.ReactNode[], number] {
+  const sections: React.ReactNode[] = [];
+  let curSectionLevel = 0;
+  let curSectionParas: React.ReactNode[] = [];
+  for (let idx = startIdx; idx < nodes.length; idx ++) {
+    const node = nodes[idx];
+    const level = getHeadingLevel(node);
+    if (level < 0) {
+      // not a heading tag
+      curSectionParas.push(node);
+    }
+    else if (level >= minLevel) {
+      if (curSectionLevel === 0 || level <= curSectionLevel) {
+        // new section
+        pushSection(curSectionParas, curSectionLevel);
+        curSectionLevel = level;
+        curSectionParas = [node];
+      }
+      else {
+        // new subsection
+        const [subsections, endIdx] = groupSections(nodes, idx, level);
+        idx = endIdx;
+        curSectionParas = [
+          ...curSectionParas,
+          ...subsections
+        ];
+      }
+    }
+    else { // level > 0 and level < minLevel
+      pushSection(curSectionParas, curSectionLevel);
+      return [sections, idx - 1];
+    }
+  }
+  pushSection(curSectionParas, curSectionLevel);
+  return [sections, nodes.length];
+
+  function pushSection(sectionParas: React.ReactNode[], level: number) {
+    if (sectionParas.length > 0) {
+      sections.push(
+        <Collapsable.Section
+          key={`section-${startIdx}-${sections.length}-${level}`}
+          level={level}>
+          {({onLoad}: {onLoad: () => void}) => {
+            onLoad();
+            return sectionParas;
+          }}
+        </Collapsable.Section>
+      );
+    }
+  }
+}
+
+interface RootWrapperProps {
+  /** Child nodes generated by markdown renderer. */
+  children: React.ReactElement[];
+}
+
+/**
+ * Render the root of markdown output by wrapping sections into collapsable
+ * containers.
+ */
+export default function renderRoot({children}: RootWrapperProps) {
+  children = groupSections(children)[0] as React.ReactElement[];
+  return <>{children}</>;
+}
